# Fix the most obvious performance pessimization caused by malice detection